+++ packages/core/event.js
@@ -23,6 +23,7 @@ class Event {
     this.app = {}
     this.device = {}
     this.request = {}
+    this.response = {}
 
     this.breadcrumbs = []
     this.threads = []
@@ -120,6 +121,7 @@ class Event {
       app: this.app,
       device: this.device,
       request: this.request,
+      response: this.response,
       breadcrumbs: this.breadcrumbs,
       context: this.context,
       groupingHash: this.groupingHash,