+++ packages/core/types/common.d.ts
@@ -120,6 +120,13 @@ interface Request {
   [key: string]: any
 }
 
+interface Response {
+  statusCode: number
+  headers: { [key: string]: unknown }
+  body?: string
+  bodyLength?: number
+}
+
 export interface User {
   id?: string
   email?: string