+++ test/browser/features/fixtures/plugin_react/webpack4/webpack.config.js
@@ -24,6 +24,6 @@ module.exports = {
     ]
   },
   plugins: [
-    new UglifyJsPlugin({ sourceMap: true, uglifyOptions: { compress: false, mangle: false, ie8: true } })
+    new UglifyJsPlugin({ sourceMap: true, uglifyOptions: { compress: false, mangle: false } })
   ]
 }