+++ test/browser/features/steps/browser_steps.rb
@@ -20,18 +20,6 @@ When('the exception matches the {string} values for the current browser') do |fi
   end
 end
 
-When('the test should run in this browser') do
-  wait = Selenium::WebDriver::Wait.new(timeout: 10)
-  wait.until {
-    Maze.driver.find_element(id: 'bugsnag-test-should-run') &&
-        Maze.driver.find_element(id: 'bugsnag-test-should-run').text != 'PENDING'
-  }
-  if Maze.driver.find_element(id: 'bugsnag-test-should-run').text == 'NO'
-    Maze::Server.reset!
-    skip_this_scenario
-  end
-end
-
 When('I let the test page run for up to {int} seconds') do |n|
   wait = Selenium::WebDriver::Wait.new(timeout: n)
   wait.until {
@@ -64,4 +52,16 @@ When('the following sets are present in the current {word} payloads:') do |reque
     Maze.check.true(payload_values.include?(expected_data),
                     "#{expected_data} was not found in any of the current payloads")
   end
+end
+
+# construct a value with placeholder using <> syntax and store it for later use
+# the placeholders are replaced with existing stored values, i.e. browser.hostname
+When('I define {string} as {string}') do |key, value|
+  resolved_value = value.gsub(/<([^>]+)>/) do
+    placeholder = $1
+    stored_value = Maze::Store.values[placeholder]
+    raise "No stored value found for placeholder '#{placeholder}'" unless stored_value
+    stored_value
+  end
+  Maze::Store.values[key] = resolved_value
 end